Move some common functions.
to keep thing dry
Fix some spelling
