+++ custom_components/ams/parsers/kaifa.py
@@ -11,6 +11,7 @@ from crccheck.crc import CrcX25
 from ..const import (DATA_FLAG, FRAME_FLAG, LIST_TYPE_LONG_1PH,
                      LIST_TYPE_LONG_3PH, LIST_TYPE_MINI, LIST_TYPE_SHORT_1PH,
                      LIST_TYPE_SHORT_3PH, WEEKDAY_MAPPING)
+from . import byte_decode, field_type
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -342,29 +343,6 @@ def parse_data(stored, data):
                 }
     stored.update(sensor_data)
     return stored, han_data
-
-
-def field_type(default="", fields=None, enc=str, dec=None):
-    """Obis/data field decoder/encoder."""
-    data = default.join(enc(i) for i in fields)
-    if dec:
-        return dec(data)
-    return data
-
-
-def byte_decode(fields=None, count=4):
-    """Data content decoder."""
-    _LOGGER.debug('fields= %s', fields)
-    if count == 2:
-        data = (fields[0] << 8 | fields[1])
-        return data
-
-    data = (fields[0] << 24 |
-            fields[1] << 16 |
-            fields[2] << 8 |
-            fields[3])
-
-    return data
 
 
 def test_valid_data(data):
