+++ custom_components/ams/parsers/kamstrup.py
@@ -9,6 +9,7 @@ from datetime import datetime
 from crccheck.crc import CrcX25
 
 from ..const import DATA_FLAG, FRAME_FLAG, WEEKDAY_MAPPING
+from . import byte_decode, field_type
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -391,29 +392,6 @@ def parse_data(stored, data):
 
     stored.update(sensor_data)
     return stored, han_data
-
-
-def field_type(default="", fields=None, enc=str, dec=None):
-    """Obis/data field decoder/encoder."""
-    data = default.join(enc(i) for i in fields)
-    if dec:
-        return dec(data)
-    return data
-
-
-def byte_decode(fields=None, count=4):
-    """Data content decoder."""
-    _LOGGER.debug('fields= %s', fields)
-    if count == 2:
-        data = (fields[0] << 8 | fields[1])
-        return data
-
-    data = (fields[0] << 24 |
-            fields[1] << 16 |
-            fields[2] << 8 |
-            fields[3])
-
-    return data
 
 
 def test_valid_data(data):
